+++ tests/testthat/test-confint-methods.R
@@ -1,5 +1,7 @@
 ## Test confint() methods
 
+var_nms <- c(".estimate", ".se", ".crit", ".lower_ci", ".upper_ci")
+
 ## first derivatives of all smooths...
 test_that("Point-wise confidence interval for a first derivatives of a GAM works", {
     skip_if_not_installed("withr")
@@ -32,8 +34,7 @@ test_that("Point-wise confidence interval for a GAM works", {
     ci <- confint(m_gam, parm = "s(x1)", type = "confidence")
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x1", "est", "se",
-                                  "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x1", var_nms))
 })
 
 test_that("Simultaneous interval for a GAM works", {
@@ -43,8 +44,7 @@ test_that("Simultaneous interval for a GAM works", {
         confint(m_gam, parm = "s(x1)", type = "simultaneous", nsim = 100))
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x1", "est", "se",
-                                  "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x1", var_nms))
 })
 
 ## 2d smooth
@@ -54,8 +54,7 @@ test_that("Point-wise confidence interval for a 2d smooth works", {
     ci <- confint(su_m_bivar_te, parm = "te(x,z)", type = "confidence")
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x", "z", "est",
-                                  "se", "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x", "z", var_nms))
 })
 
 test_that("Simultaneous interval for a 2d smooth works", {
@@ -66,8 +65,7 @@ test_that("Simultaneous interval for a 2d smooth works", {
             nsim = 100))
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x", "z", "est",
-                                  "se", "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x", "z", var_nms))
 })
 
 test_that("Point-wise confidence interval for a GAMM works", {
@@ -76,8 +74,7 @@ test_that("Point-wise confidence interval for a GAMM works", {
     ci <- confint(m_gamm, parm = "s(x1)", type = "confidence")
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x1", "est", "se",
-                                  "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x1", var_nms))
 })
 
 test_that("Simultaneous interval for a GAMM works", {
@@ -87,8 +84,7 @@ test_that("Simultaneous interval for a GAMM works", {
         confint(m_gamm, parm = "s(x1)", type = "simultaneous", nsim = 100))
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x1", "est", "se",
-                                  "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x1", var_nms))
 })
 
 ## confint methods for by variables
@@ -99,8 +95,7 @@ test_that("Point-wise confidence interval for a GAM with factor by variable work
                   partial_match = TRUE)
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x2", "fac", "est",
-                                  "se", "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x2", "fac", var_nms))
     expect_equal(paste0("s(x2):fac", levels(su_eg4[["fac"]])),
                  unique(ci[["smooth"]]))
 })
@@ -112,8 +107,7 @@ test_that("Simultaneous confidence interval for a GAM with factor by variable wo
                   partial_match = TRUE)
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x2", "fac", "est",
-                                  "se", "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x2", "fac", var_nms))
     expect_equal(paste0("s(x2):fac", levels(su_eg4[["fac"]])),
                  unique(ci[["smooth"]]))
 })
@@ -125,8 +119,7 @@ test_that("Point-wise confidence interval for a GAM with selected factor by vari
     ci <- confint(su_m_factor_by_x2, parm = "s(x2):fac1", type = "confidence")
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x2", "fac", "est",
-                                  "se", "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x2", "fac", var_nms))
 })
 
 test_that("Point-wise confidence interval for a GAMM works", {
@@ -135,8 +128,7 @@ test_that("Point-wise confidence interval for a GAMM works", {
     ci <- confint(m_gamm4, parm = "s(x1)", type = "confidence")
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x1", "est", "se",
-                                  "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x1", var_nms))
 })
 
 test_that("Simultaneous interval for a GAMM works", {
@@ -146,8 +138,7 @@ test_that("Simultaneous interval for a GAMM works", {
         confint(m_gamm4, parm = "s(x1)", type = "simultaneous", nsim = 100))
     expect_s3_class(ci, "confint.gam")
     expect_s3_class(ci, "tbl_df")
-    expect_named(ci, expected = c("smooth", "type", "by", "x1", "est", "se",
-                                  "crit", "lower", "upper"))
+    expect_named(ci, expected = c("smooth", "type", "by", "x1", var_nms))
 })
 
 ## test snapshots...